introduce ArtifactManager for artifact handling
