+++ vedro/plugins/artifacted/__init__.py
@@ -1,5 +1,6 @@
 from vedro.core import Artifact, FileArtifact, MemoryArtifact
 
+from ._artifact_manager import ArtifactManager
 from ._artifacted import (
     Artifacted,
     ArtifactedPlugin,
@@ -11,4 +12,4 @@ from ._artifacted import (
 
 __all__ = ("Artifacted", "ArtifactedPlugin", "attach_artifact", "attach_step_artifact",
            "attach_scenario_artifact", "attach_global_artifact", "Artifact",
-           "MemoryArtifact", "FileArtifact",)
+           "MemoryArtifact", "FileArtifact", "ArtifactManager",)
